requirements.txt
dask==2024.5.1
Flask==3.1.1
joblib==1.5.0
numpy==1.26.4
pandas==2.3.1
python-dotenv==1.1.1
scikit_learn==1.7.1
scipy==1.15.3
shap==0.48.0
tmdbv3api==1.9.0
tqdm==4.67.1
Werkzeug==3.1.3
xgboost==3.0.4